rename problem set group modal
