+++ frontend/src/pages/admin/components/ProblemSetGroupModal.vue
